+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/DbUserStorage.java
@@ -1,12 +1,14 @@
 package ru.yandex.practicum.filmorate.storage.user;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.AlreadyExistException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
@@ -20,6 +22,7 @@ import java.util.List;
 import java.util.Objects;
 
 @Component
+@Slf4j
 @RequiredArgsConstructor
 public class DbUserStorage implements UserStorage{
     private final JdbcTemplate jdbcTemplate;
@@ -82,15 +85,21 @@ public class DbUserStorage implements UserStorage{
     }
 
     @Override
-    public User deleteById(int id) {
-        final String sqlQuery = "DELETE FROM USERS WHERE USER_ID = ?";
-        try {
-            jdbcTemplate.update(sqlQuery, id);
-        } catch (Exception e) {
-            System.out.println(e);
+    public boolean deleteById(int id) {
+
+        String checkQuery = "SELECT * FROM USERS WHERE USER_ID = ?";
+        SqlRowSet filmRows = jdbcTemplate.queryForRowSet(checkQuery, id);
+        if (!filmRows.next()) {
+            log.warn("Пользователь с идентификатором {} не найден.", id);
+            throw new NotFoundException("Пользователь не найден");
         }
-        User user = getUser(id);
-        return user;
+
+        final String sqlQuery = "DELETE FROM USERS WHERE USER_ID = ?";
+
+         jdbcTemplate.update(sqlQuery, id);
+
+     //   User user = getUser(id);
+        return true;
     }
 
     @Override
@@ -123,6 +132,14 @@ public class DbUserStorage implements UserStorage{
 
     @Override
     public List<User> getUserFriends(int userId) {
+
+        String checkQuery = "SELECT * FROM USERS WHERE USER_ID = ?";
+        SqlRowSet filmRows = jdbcTemplate.queryForRowSet(checkQuery, userId);
+        if (!filmRows.next()) {
+            log.warn("Пользователь с идентификатором {} не найден.", userId);
+            throw new NotFoundException("Пользователь не найден");
+        }
+
         final String sqlQuery = "SELECT * " +
                 "FROM USERS " +
                 "WHERE USER_ID IN (" +
@@ -157,3 +174,8 @@ public class DbUserStorage implements UserStorage{
                 .build();
     }
 }
+
+/*
+        try {       } catch (Exception e) {
+            System.out.println(e);
+        } */
